Introduce subscriber directory
Formerly, we were subscribing to topics by directly providing a
callback. This had the drawback that, in the case of anonymous
lambdas, we may not be able to identify distinct subscribers, and so
could not deduplicate them nor reliably support unsubscribing from
topics.

This introduces a "directory" of subscribers to complement the "board"
of topics. The subscribers are registered with both a name and a
callback, and they are added to the topics just by name.